Add start script to package.json, adjust session middleware order, and enhance layout middleware to merge options

+++ src/utils/index.js
@@ -8,16 +8,38 @@ const getNav = async () => {
     const name = row.category_name;
     nav += `<li><a href="/category/view/${id}">${name}</a></li>`;
   });
-  return `
-    ${nav}
+
+  if (user) {
+    nav += `
         <li><a href="/game/add">Add Game</a></li>
         <li><a href="/category/add">Add Category</a></li>
         <li><a href="/category/delete">Delete Category</a></li>
+        <li><a href="/About">About Me</a></li>
+        <li><a href="/account/logout">Logout</a></li>
+        </ul>
+    </nav>`;
+    return nav;
+  } else {
+    nav += `
         <li><a href="/About">About Me</a></li>
         <li><a href="/account/login">Login</a></li>
         <li></li><a href="/account/register">Register</a></li>
         </ul>
     </nav>`;
+    return nav;
+  }
+
+  return nav;
+  // return `
+  //   ${nav}
+  //       <li><a href="/game/add">Add Game</a></li>
+  //       <li><a href="/category/add">Add Category</a></li>
+  //       <li><a href="/category/delete">Delete Category</a></li>
+  //       <li><a href="/About">About Me</a></li>
+  //       <li><a href="/account/login">Login</a></li>
+  //       <li></li><a href="/account/register">Register</a></li>
+  //       </ul>
+  //   </nav>`;
 };
 
 const requireAuth = (req, res, next) => {
